app router - /reschedule pages
