+++ apps/web/app/reschedule/[uid]/embed/page.tsx
@@ -1,23 +1,18 @@
-import { withAppDirSsr } from "app/WithAppDirSsr";
 import type { PageProps } from "app/_types";
-import { _generateMetadata } from "app/_utils";
-import { headers, cookies } from "next/headers";
+import { cookies, headers } from "next/headers";
 
 import { buildLegacyCtx } from "@lib/buildLegacyCtx";
 import { getServerSideProps } from "@lib/reschedule/[uid]/getServerSideProps";
+import withEmbedSsrAppDir from "app/WithEmbedSSR";
+import { withAppDirSsr } from "app/WithAppDirSsr";
 
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    () => "",
-    () => ""
-  );
 
 const getData = withAppDirSsr(getServerSideProps);
+const getEmbedData = withEmbedSsrAppDir(getData);
 
 const Page = async ({ params, searchParams }: PageProps) => {
   const legacyCtx = buildLegacyCtx(headers(), cookies(), params, searchParams);
-
-  await getData(legacyCtx);
+  await getEmbedData(legacyCtx);
 
   return null;
 };
