+++ apps/web/app/reschedule/[uid]/page.tsx
@@ -1,0 +1,25 @@
+import { withAppDirSsr } from "app/WithAppDirSsr";
+import type { PageProps } from "app/_types";
+import { _generateMetadata } from "app/_utils";
+import { headers, cookies } from "next/headers";
+
+import { buildLegacyCtx } from "@lib/buildLegacyCtx";
+import { getServerSideProps } from "@lib/reschedule/[uid]/getServerSideProps";
+
+export const generateMetadata = async () =>
+  await _generateMetadata(
+    (t) => t("reschedule"),
+    () => ""
+  );
+
+const getData = withAppDirSsr(getServerSideProps);
+
+const Page = async ({ params, searchParams }: PageProps) => {
+  const legacyCtx = buildLegacyCtx(headers(), cookies(), params, searchParams);
+
+  await getData(legacyCtx);
+
+  return null;
+};
+
+export default Page;
