--- apps/web/pages/reschedule/[uid].tsx
@@ -1,6 +1,0 @@
-export default function Type() {
-  // Just redirect to the schedule page to reschedule it.
-  return null;
-}
-
-export { getServerSideProps } from "@lib/reschedule/[uid]/getServerSideProps";
