--- apps/web/pages/reschedule/[uid]/embed.tsx
@@ -1,7 +1,0 @@
-import withEmbedSsr from "@lib/withEmbedSsr";
-
-import { getServerSideProps as _getServerSideProps } from "../[uid]";
-
-export { default } from "../[uid]";
-
-export const getServerSideProps = withEmbedSsr(_getServerSideProps);
